Fix --separate-debug-symbols
Previously cargo-deb attempted to add the files with debug symbols by
referencing the stripped executable path.  The debug symbols do not live
alongside the stripped executables so this always failed.

Now cargo-deb uses the original executable paths to determine the
location of the debug symbols.

(There's probably a better way to do this, but I'm not skilled enough to
figure out how to use collect() with the parallel iterator.)