Allow depends, pre-depends, recommends, and suggests to be arrays
This commit makes it so that depends (and any of the other fields above)
can be written as

   depends = ["a", "b", "c"]

in addition to the existing depends = "a, b, c" format.

It is converted at the config level by joining all the array elements
with ", ". No additional validation is performed beyond that already
done by serde deserialization.

Closes #88