Use `zopfli` instead of `flate2` for Gzip compression
As mentioned in #100, `cargo-deb` used to use the Zopfli crate for
compressing data using the Gzip format, which provides overall smaller
results than other general-purpose Gzip encoders that cater to speed
more. I think this is a good tradeoff for data meant to be compressed
once and distributed or decompressed many times, such as Debian
packages.

However, the `zopfli` crate API did not support the streaming Gzip
encoding that `cargo-deb` moved to. This has changed with its latest
v0.8.0 release, bringing it more on par with `flate2`, so that's no
longer a reason to not use Zopfli.

I ran the available test suite and noticed no regressions due to this
change.