Use separate-debug-symbols value from Cargo.toml
Previously this was documented to be configurable from the
[[package.metadata.deb]] table but was in fact ignored and
only the command line parameter was used.

This honors the value from Cargo.toml if no command line option
is specified. The existing --separate-debug-symbols option can
still be used to enable the inclusion of a separate debug symbol
file. A new --no-separate-debug-symbols option can be used to
suppress it, overriding a possible Cargo.toml value that would have
led to its inclusion.