Remove the sha256 file.
lintian gives a warning if you have the sha256 file present and a
sha does not exist for the binary control file. We leave the
funtion around because if you ever impliment the source control
file or .changes file it does exist.

We are not updating the CHANGELOG.md file for fear of a conflict.

Closes #145